+++ frontend/connection_tab_widget.py
@@ -1,4 +1,4 @@
-from PyQt6.QtWidgets import (QWidget, QLabel,
+from PyQt6.QtWidgets import (QLabel,
                              QLineEdit, QSpinBox, QVBoxLayout)
 
 from frontend.base_detail_widget import BaseRequest
@@ -41,7 +41,7 @@ class ModbusTcpConnectionGrid(BaseConnectionGrid):
 
         self.retries_spinbox = QSpinBox()
         self.retries_spinbox.setRange(1, 65535)
-        self.grid_layout.add_widget(QLabel("Timeout:"), self.retries_spinbox)
+        self.grid_layout.add_widget(QLabel("Retries:"), self.retries_spinbox)
 
         # Set initial state and connect signals:
         self.update_view(data=self.item_client)
@@ -55,19 +55,18 @@ class ModbusTcpConnectionGrid(BaseConnectionGrid):
                 "timeout": int(self.timeout_spinbox.text()),
                 "retries": int(self.retries_spinbox.text()),
             }
-            self.api_client.update_item_from_handler(
-                item_handler=self.item_client["item_handler"],
-                item_id=self.item_client["item_id"],
-                **client,
-                callback=self.update_view,
-            )
+
+            self.call_api(api_method="update_item_from_handler",
+                          item_handler=self.item_client["item_handler"],
+                          item_id=self.item_client["item_id"],
+                          **client,
+                          callback=self.update_view)
         else:
-            self.api_client.create_client_item(
-                item_name=self.item["name"],
-                item_handler="ModbusTcpClient",
-                parent_item_id=self.item["item_id"],
-                callback=self.update_view,
-            )
+            self.call_api(api_method="create_client_item",
+                          item_name=self.item["name"],
+                          item_handler="ModbusTcpClient",
+                          parent_item_id=self.item["item_id"],
+                          callback=self.update_view)
 
     def update_view(self, data: dict):
         if not data:
@@ -133,19 +132,18 @@ class ModbusRtuConnectionGrid(BaseConnectionGrid):
                 "timeout": int(self.timeout_spinbox.text()),
                 "retries": int(self.retries_spinbox.text()),
             }
-            self.api_client.update_item_from_handler(
-                item_handler=self.item_client["item_handler"],
-                item_id=self.item_client["item_id"],
-                **client,
-                callback=self.update_view,
-            )
+
+            self.call_api(api_method="update_item_from_handler",
+                          item_handler=self.item_client["item_handler"],
+                          item_id=self.item_client["item_id"],
+                          **client,
+                          callback=self.update_view)
         else:
-            self.api_client.create_client_item(
-                item_name=self.item["name"],
-                item_handler="ModbusRtuClient",
-                parent_item_id=self.item["item_id"],
-                callback=self.update_view,
-            )
+            self.call_api(api_method="create_client_item",
+                          item_name=self.item["name"],
+                          item_handler="ModbusRtuClient",
+                          parent_item_id=self.item["item_id"],
+                          callback=self.update_view)
 
     def update_view(self, data: dict):
         if not data:
@@ -203,14 +201,13 @@ class ConnectionTabWidget(BaseRequest):
     def update_item(self):
         new_connection_type = self.connection_type_combo.currentText()
         if self.item["client_type"] != new_connection_type:
-            self.api_client.update_item_from_handler(
-                item_handler=self.item["item_handler"],
-                item_id=self.item["item_id"],
-                client_id=None,
-                client_type=new_connection_type,
-                client=None,
-                callback=self.update_view
-            )
+            self.call_api(api_method="update_item_from_handler",
+                          item_handler=self.item["item_handler"],
+                          item_id=self.item["item_id"],
+                          client_id=None,
+                          client_type=new_connection_type,
+                          client=None,
+                          callback=self.update_view)
 
     def update_view(self, data: dict = None):
         """Switch between GridLayouts when QComboBox changes value."""