+++ frontend/main_window.py
@@ -1,5 +1,6 @@
 import argparse
 import sys
+import traceback
 
 from PyQt6.QtGui import QIcon
 from PyQt6.QtWidgets import (
@@ -10,17 +11,19 @@ from PyQt6.QtWidgets import (
     QVBoxLayout,
     QWidget,
     QPushButton,
-    QSizePolicy
+    QSizePolicy, QMessageBox
 )
 
-from frontend.api_client import ApiClient
+from frontend.api.api_helper_mixin import ApiCallMixin
+from frontend.api.api_client import ApiClient
 from frontend.collection_detail_widget import CollectionDetail
 from frontend.common import ITEMS
+from frontend.menu_widget import AppInfo
 from frontend.project_structure_section import ProjectStructureSection
 from qt_material import apply_stylesheet
 from frontend.modbus_detail_widget import ModbusDetail
 
-from config import FRONTEND_PATH, load_app_config
+from config import FRONTEND_PATH, load_app_config, PROJECT_DATA_PATH
 
 
 class Button(QPushButton):
@@ -32,7 +35,7 @@ class Button(QPushButton):
         self.adjustSize()
 
 
-class MainWindow(QMainWindow):
+class MainWindow(QMainWindow, ApiCallMixin):
     def __init__(self, host: str, port: int):
         super().__init__()
 
@@ -43,6 +46,11 @@ class MainWindow(QMainWindow):
 
         # Set API client:
         self.api_client = ApiClient(host=host, port=port)
+        self.setup_api_client(self.api_client)
+
+        # Setup menu
+        self.app_info = AppInfo()
+        self.setup_menu()
 
         # Divide window in two sections:
         self.main_window_sections_splitter = QSplitter()
@@ -71,10 +79,32 @@ class MainWindow(QMainWindow):
         container.setLayout(self.main_window_layout)
         self.setCentralWidget(container)
 
+    def setup_menu(self):
+        """Create the application menu bar"""
+        menubar = self.menuBar()
+
+        # Help menu
+        help_menu = menubar.addMenu("&Help")
+
+        # Add actions
+        help_menu.addAction("Logs Path", lambda: QMessageBox.information(
+            self,
+            "Logs Location",
+            f"Logs are stored at:\n{PROJECT_DATA_PATH}"
+        ))
+        help_menu.addSeparator()
+        help_menu.addAction("License", lambda: self.app_info.show_license(self))
+        help_menu.addAction("Documentation", lambda: self.app_info.show_readme(self))
+        help_menu.addAction("About", lambda: self.app_info.show_about(self))
+
     def get_item_request(self):
+        print("get_item_request")
         item_id = self.project_structure_section.get_selected_item_data()
+        print("item_id", item_id)
         if item_id:
-            self.api_client.get_item_request(item_id, callback=self.set_detail_section)
+            self.call_api(api_method="get_item_request",
+                          item_id=item_id,
+                          callback=self.set_detail_section)
 
     def set_detail_section(self, item = None, *args, **kwargs):
         if item is not None:
@@ -95,7 +125,8 @@ class MainWindow(QMainWindow):
     def closeEvent(self, *args, **kwargs):
         """Override the close event to perform custom actions."""
         # Wait until backend stops:
-        self.api_client.stop_item(item_id=0)
+        self.call_api(api_method="stop_item",
+                      item_id=0)
 
         super().closeEvent(*args, **kwargs)
 
@@ -104,10 +135,14 @@ def run(host: str, port: int):
     app = QApplication(sys.argv)
     apply_stylesheet(app, theme=f"{FRONTEND_PATH}/fixtures/theme.xml", css_file=f"{FRONTEND_PATH}/fixtures/styles.css")
 
-    window = MainWindow(host=host, port=port)
-    window.show()
+    try:
+        window = MainWindow(host=host, port=port)
+        window.show()
 
-    sys.exit(app.exec())
+        exec_obj = app.exec()
+        sys.exit(exec_obj)
+    except Exception as e:
+        print(f"Main window critical error: {str(e)}: {traceback.format_exc()}")
 
 
 if __name__ == "__main__":